pyproject.toml
[project]
name = "discord-flux-bridgeremake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2026.1.4",
    "discord-py>=2.6.4",
    "fluxer-py>=0.1.2",
    "python-dotenv>=1.2.1",
]
